Se ha corregido la interfaz

+++ app/Http/Controllers/FolderController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Folder;
 use App\Models\Departament;
+use Illuminate\Support\Facades\Auth;
 
 class FolderController extends Controller
 {
@@ -12,7 +13,7 @@ class FolderController extends Controller
     {
 
         $folders=\DB::table('folders AS d1')
-   
+   ->where('d1.departament_id','=',Auth::user()->departament_id)
         ->join('folders AS d2','d2.id','=','d1.father_folder_id')
     ->join('departaments AS d3','d3.id','=','d1.departament_id')
     ->select('d1.*', 'd2.name as father_folder', 'd3.name as departament')
@@ -38,19 +39,31 @@ class FolderController extends Controller
         $this->validate($request, $rules, $messages);
         $folders= new Folder();
         $folders->name= $request->input('name');
-        $folders->departament_id= $request->input('departament');
-        $folders->father_folder_id= $request->input('padre');
+        $folders->departament_id= Auth::user()->departament_id;
+        if($request->input('padre')==null)
+        {
+            $folders->father_folder_id= 1;
+        }else{
+            $folders->father_folder_id= $request->input('padre');
+        }
+    
         $folders->save();
  
         return back()->with('notification','La carpeta ha sido registrado exitosamente');
     }
     public function edit($id)
     {
+
         $folder=Folder::find($id);
+        if($folder->departament_id!=Auth::user()->departament_id){
+return('Usted no tiene permisos para realizar esta operacion');
+        }
         $departaments=Departament::all();
-        $father_folders=Folder::all();
+        $father_folders=\DB::table('folders AS d1')
+        ->where('d1.departament_id','=',Auth::user()->departament_id)->get();
         return view('admin.folders.edit')->with(compact('folder'))->with(compact('departaments'))->with(compact('father_folders'));
     }
+
     public function update($id, Request $request)
     {
         $rules = [
@@ -66,7 +79,7 @@ class FolderController extends Controller
         $this->validate($request, $rules, $messages);
         $folders=  Folder::find($id);
         $folders->name= $request->input('name');
-        $folders->departament_id= $request->input('departament');
+        
         $folders->father_folder_id= $request->input('padre');
         $folders->save();
         
@@ -74,6 +87,9 @@ class FolderController extends Controller
     }
     public function delete($id){
         $folder =Folder::find($id);
+        if($folder->departament_id!=Auth::user()->departament_id){
+            return('Usted no tiene permisos para realizar esta operacion');
+                    }
         $folder->delete();
         return back()->with('notification','La carpeta ha sido dado de baja exitosamente');
            }
